move concatenate fn to xebec
