+++ xebec/__init__.py
@@ -1,4 +1,7 @@
+import os
+
 from .src.helper import get_logger
 
 
 __version__ = "0.3.0"
+COOKIE_DIR = os.path.dirname(__file__)
