+++ xebec/{{cookiecutter.project_name}}/workflow/scripts/concatenate.py
@@ -3,33 +3,7 @@ from pathlib import PurePath
 import pandas as pd
 
 from xebec import get_logger
-
-
-def concatenate_metric_dataframes(files):
-    """Concatenate results from multiple metrics."""
-    def get_metric_info(f):
-        """Return metric type and metric name as tuple.
-
-        results/*_div/*/*/effect_sizes.tsv
-        """
-        path_parts = PurePath(f).parts
-        return path_parts[2], path_parts[3]
-
-    all_dfs = []
-    all_keys = []
-    for f in files:
-        this_df = pd.read_table(f, sep="\t")
-        this_keys = get_metric_info(f)
-        all_dfs.append(this_df)
-        all_keys.append(this_keys)
-
-    total_df = pd.concat(
-        all_dfs,
-        keys=all_keys,
-        names=["phylogenetic", "diversity_metric"]
-    )
-    total_df = total_df.reset_index(level=("phylogenetic", "diversity_metric"))
-    return total_df
+from xebec.src._concatenate import concatenate_metric_dataframes
 
 
 xebec_logger = get_logger(snakemake.log[0], snakemake.rule)
